requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Performance
numpy==1.26.3
python-Levenshtein==0.25.0

# Meilisearch
meilisearch-python-sdk==2.7.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Monitoring
prometheus-client==0.19.0
psutil==5.9.7

# Utils
python-dotenv==1.0.0
